pyproject.toml
[project]
name = "roz"
version = "0.1.0"
description = "Always-watching surveillance assistant with LLM scene analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "requests>=2.31",
    "kokoro>=0.9.4",
    "piper-tts>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.10",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
